final changes for now
as of now, the project is now complete. the game is fully playable (with a possibility of bugs that haven't caught my attention) and is in a good state as of now to begin my second project. this wont be the final addition tho. as I plan to in the future simplify as well improve the gameplay / interface of the game. as of now, updates are on a pause as a work on a new project

this game is over 500 lines of code non-simplified. maybe I can reduce it in the future and make separate files for everything. as of now, it is one big mess